Ideally we should be able to share React.cache but we cannot yet do it.
I tried to call the action as a component but that ends up yielding other issues like the result being converted to React.lazy or debug info being lost.

It's also wrong to render it as a component because it would become the owner where as the owner should be null.

We really need a way to expand the React.cache scope.