[form] allow turning prefetching off (#68305)
Adds a `prefetch` prop to Form. It mirrors how Link's `prefetch` works,
but it only allows `prefetch={false}` (no prefetching) and
`prefetch={null}` ("auto" prefetch, the default). `prefetch={true}` does
not make sense for Form.

I don't expect this to be used much, because prefetching is one of the
main parts of Form's functionality, but i think it's good to have the
option to disable prefetching (even if mostly as an escape hatch)