Unwrap `registerServerReference` function
In the same vein as #69190, where we already unwrapped
`createServerReference`, we now also unwrap `registerServerReference`,
which is required for React to select the right call stack frame when
generating source locations for server actions facebook/react#30741.

Whereas unwrapping of `createServerReference` was required for server
actions that are imported into client components, unwrapping
`registerServerReference` is needed for server actions that are passed
from server components to client components.

This does not fully enable the source mapping just yet. For this to work
end-to-end, the next step is to generate proper spans in the SWC
transform, which will be done in the next PR.