Exclude Object and Promise prototype properties from shadowing for params and searchParams
params and searchParams are now promises however to facilitate migration params and searchParams can still be referenced directly on these props. There are a number of special properties however that conflict with this and this change special cases more property names to not be synchronously accessed.

We exclude common Object prototype properties, Promise prototype properties, and properties that are commonly existence tested like toJSON and displayName.