next-codemod: Ensure install packages works on exotic package revisions
Spawning a process with multiple args is not the same as spawning it with
a single arg that was created from args.join(' ').
This works most of the time but broke when I tried to use specific protocols e.g. `@types/react@npm:types-react@rc`.
This is a valid revision but not parsed properly if passed as a single arg to the dependencies.