Error consistently during prospective prerender
When `dynamicIO` is enabled there is a prospective render used to fill caches. Previously we ignored errors during this phase. However this might mask user errors that are flakey. With this change we consider errors during the prospective render valid and throw them to the same handling logic that handles errors for the final render. We exclude control errors like prerender complete and prerender interrupted.