refactor: remove unnecessary page segment check in ppr-navigations (#70572)
In the PPR navigations flow, we were always spawning a pending task for
page segments, which seemed to be related to search params not updating.
(x-ref: #60242)

However this case doesn't seem to be necessary: the static prefetch
shouldn't contain any data related to search params, as accessing search
params will postpone and trigger suspense fallbacks. And when the
dynamic request sends down segment key information for page segments
search params, the params will be encoded in the segment key (e.g.
`__PAGE__{'foo': 'bar'}`, as opposed to just `__PAGE__`.

I can't think of another case where this handling would be necessary as
if we are in a new tree, we'll already spawn a pending task. And if
we're in the same tree, I expect the static data should be the same,
while the dynamic request will contain the different data.