next-upgrade: Allow custom selection of codemods
Instead of having to choose between all or none of the codemods, each application
can now be select up front.
We select all by default so that you still have just one keypress
for the happy path.
But sometimes codemods are added later.
Or we have codemods that may not be suitable depending on how code is structured.
Or we have optional codemods.

For me specifically it was just motivated by not being able to select a specific version yet
and skipping the async API codemod while it's under development.