next-upgrade: Default to release channel that's used
For example, `npx @next/codemod@canary` will default to the Canary release channel.
`npx @next/codemod@latest` will use the Stable release channel.

This works by reading the `version` field in the `package.json`.
If `canary` is in the version, it will default to the Canary release channel.
If `rc` is in the version, it will default to the RC release channel.
Otherwise we use `latest`.