Fix interpolation of dynamic params in intercepting route pathnames
In a pathname for an intercepting route, a dynamic path param might
occur multiple types, e.g.
`'/[locale]/example/(...)[locale]/intercepted'`. We need to make sure
that each occurence is replaced with the actual value during
interpolation.

The handling of such pathnames is already done correctly in the Node.js
runtime, but not in the Edge runtime.

fixes #70654