initially visited static pages should respect static staletime (#70640)
This provides information to the initial SSR render that the client
leverages when seeding the prefetch cache, to determine if a page was a
fully static prerender.

The side effect of not doing this is that if we seeded a static page
during SSR, and then navigated to it, the router would make a request
for data that it didn't need. In other words, it didn't honor the static
staletime, like other navigations would.