Ensure that client reference module export names are not mangled
This is required for client reference modules in the RSC layer to be
resolvable by React using the metadata from the client manifest.

Using `setUsedInUnknownWay` to tell Webpack that exports are used in an
unknown way is preferable over injecting runtime code that accesses the
exports dynamically.

The change is covered by the test `use-cache > should cache results`
(`pnpm test-start test/e2e/app-dir/use-cache/use-cache.test.ts`), which
currently fails in #70524.

As a follow-up, we need to do something similar in Turbopack as well.