tests: deprecate `check` & disable flaky app-action deploy test (#70726)
We've discussed internally deprecating `check` so I've marked it as
deprecated in this PR and refactored app-action to use `retry` +
`expect` in favor of `check`. `check` makes it really hard to tell what
the expected/received behavior is and it becomes hard to reason about
when we're asserting on strings that return `'success'` or `'yes'`

Other misc improvement:
[improve middleware rewrite
test](https://github.com/vercel/next.js/pull/70726/commits/1cc3cf3aecf5975f8bf78e7523901b2ca57d3b80)

I also disabled the revalidateTag test as it's consistently flaking when
deployed and I haven't been able to diagnose what's wrong with it.