[Turbopack] add new backend (#69667)
### What?

Add a new turbo-tasks backend that stores data normalized and is
prepared for persistent caching.

No longer keeps state in task locals, but all state is stored
immediately. This is required for persistent caching since it could
store and resume at any point and we always need to be in a valid state.

Tasks stay dirty until recomputed. (Previously dirty was changed to
scheduled)

### Why?

* Reduced memory usage
* preparation for persistent caching