Update ga.tsxUpdate ga.tsx to support standalone usage of sendGAEvent
The current implementation prevents the sendGAEvent function from being used if Google Tag Manager is imported with a custom script and not with the GoogleAnalytics function, this is because the currDataLayerName variable is initialized to undefined and initialized to “dataLayer” (or a string passed in as prop) only in the GoogleAnalytics function and, in the sendGAEvent function, it is checked to see if that value is undefined and, if so, a warning is triggered. By using the sendGAEvent function standalone, it is impossible for the currDataLayerName variable to be initialized and so the warning always starts, preventing its use.