Tracing: Always end write stream on shutdown in dev
While we don’t want to end the write stream in other situations in dev, we _do_ want to do this when shutting down the web server.

This previously resulted in some spans, including the top-level `next-dev` span, not always being written to disk as the process would exit before it was flushed.

Test Plan: Ensured `next-dev` span is written to trace on devserver shutdown