Defensively install patches
Some builtins and node library modules may make patching impossible by freezing objects or making propeties readonly. The patch installation is done as early as possible before we know if we're even going to need the patch and thus is done unconditionally whether DIO is enabled or not. This is generally preferable to how for instance the fetch patch is installed which is observable becasue user code can run before it.

In this change all the patch installs are for dynamicIO augmented behavior. We can allow the process to run but will warn that `experimental.dynamicIO` will not have expected behavior when calling the functions that were not installed.