Move tracking of revalidate and tags from WorkStore to WorkUnitStore (#70940)
This allows to have individually track the tags / revalidate for each
scope - such as the smallest revalidate number within a cache scope.
Instead of the whole page.

Currently, we don't use the collected revalidate/tags for a dynamic
render so the RequestStore doesn't collect them. Only the Prerender and
Cache store. Also not unstable_cache since it doesn't have nested
caches.

Instead of using `false` to mean don't cache, I use a very large number
to mean "Infinity". This is the largest SMI. I.e. 31-bit signed integer.
Which ensures that we're not deopting into Floats in V8. This number is
also serializable through JSON so if we leak it, it still makes sense.