Add Next-Router-Segment-Prefetch header
Introduces a new App Router header that will be used for prefetching
individual segments. It will accept a path to the segment that is
being requested.

The header is treated similarly to the other known headers: it is
stripped by `stripFlightHeaders` before the application handles the
request, and it's part of the Vary header so clients cache
it correctly.

This does not yet implement any of the actual response handling.