Add RequestStore around renders
Today we scope the entire work store with a RequestStore. This is because prerenders still have some dependency on this store existing. We want to get rid of this but to do so we need to ensure there is a RequestStore around the renders inside renderToStream. This change adds the necessary scoping inside renderToStream. Today this is redundant but it sets us upto be able to remove the outer RequestStore scope.

In addition I removed the withRequestStore helper. We generally don't want to construct a new RequestStore for everything we will scope in a RequestStore scope so we can just create the store and then wrap the necessary functions with workUnitAsyncStorage.run...