Export segment prefetch data during build
Implements exporting segment prefetch data during build.

Although the goal is to support individual segments being requested by
the client, all of the segments for a page are generated simultaneously,
including during revalidations. This is to ensure consistency, because
it's possible for a mismatch between a layout and page segment can
cause the client to error during rendering. We want to preserve the
ability of the client to recover from such a mismatch by re-requesting
all the segments to get a consistent view of the page.

The segments data is generated during the same phase that we currently
generate prefetch data for the entire page.

In this commit, I have not yet implemented the actual segment
generation; I'm using placeholder data for now. I've only implemented
the plumbing of the data through the build process. I will implement
the segment generation in a separate PR.